fix(高温告警)

+++ src/main/java/com/xiliulou/electricity/dto/ElectricityCabinetOtherSetting.java
@@ -141,5 +141,5 @@ public class ElectricityCabinetOtherSetting {
     /**
      * 高温告警阈值
      */
-    private Integer temperatureWarningValue;
+    private Double temperatureWarningValue;
 }
